ref(api): seperate api read/get functions into separate file
