+++ src/3box.js
@@ -14,6 +14,7 @@ const Space = require('./space')
 const utils = require('./utils/index')
 const verifier = require('./utils/verifier')
 const config = require('./config.js')
+const API = require('./api')
 
 const ADDRESS_SERVER_URL = config.address_server_url
 const PINNING_NODE = config.pinning_node
@@ -21,8 +22,6 @@ const PINNING_ROOM = config.pinning_room
 // const IFRAME_STORE_VERSION = '0.0.3'
 // const IFRAME_STORE_URL = `https://iframe.3box.io/${IFRAME_STORE_VERSION}/iframe.html`
 const IPFS_OPTIONS = config.ipfs_options
-const GRAPHQL_SERVER_URL = config.graphql_server_url
-const PROFILE_SERVER_URL = config.profile_server_url
 
 let globalIPFS, globalOrbitDB // , ipfsProxy, cacheProxy, iframeLoadedPromise
 
@@ -177,8 +176,7 @@ class Box {
     opts = Object.assign({ useCacheService: true }, opts)
     let profile
     if (opts.useCacheService) {
-      const profileServerUrl = opts.profileServer || PROFILE_SERVER_URL
-      profile = await getProfileAPI(normalizedAddress, profileServerUrl)
+      profile = await API.getProfile(normalizedAddress, opts.profileServer)
     } else {
       profile = await this._getProfileOrbit(normalizedAddress, opts)
     }
@@ -195,15 +193,13 @@ class Box {
    */
 
   static async getProfiles (addressArray, opts = {}) {
-    const profileServerUrl = opts.profileServer || PROFILE_SERVER_URL
-    const req = { addressList: addressArray }
-    return utils.fetchJson(profileServerUrl + '/profileList', req)
+    return API.getProfiles(addressArray, opts)
   }
 
   static async _getProfileOrbit (address, opts = {}) {
     // opts = Object.assign({ iframeStore: true }, opts)
-    const serverUrl = opts.addressServer || ADDRESS_SERVER_URL
-    const rootStoreAddress = await getRootStoreAddress(serverUrl, address.toLowerCase())
+    console.log(opts.addressServer)
+    const rootStoreAddress = await API.getRootStoreAddress(address.toLowerCase(), opts.addressServer)
     let usingGlobalIPFS = false
     let usingGlobalOrbitDB = false
     let ipfs
@@ -276,7 +272,7 @@ class Box {
    */
 
   static async profileGraphQL (query, opts = {}) {
-    return graphQLRequest(opts.graphqlServer || GRAPHQL_SERVER_URL, query)
+    return API.profileGraphQL(query, opts.graphqlServer)
   }
 
   /**
@@ -466,27 +462,6 @@ async function initIPFS (ipfs, iframeStore, ipfsOptions) {
       ipfs.on('ready', () => resolve(ipfs))
     })
   }
-}
-
-async function getRootStoreAddress (serverUrl, identifier) {
-  // read orbitdb root store address from the 3box-address-server
-  const res = await utils.fetchJson(serverUrl + '/odbAddress/' + identifier)
-  if (res.status === 'success') {
-    return res.data.rootStoreAddress
-  } else {
-    throw new Error(res.message)
-  }
-}
-
-async function getProfileAPI (rootStoreAddress, serverUrl) {
-  return new Promise(async (resolve, reject) => {
-    try {
-      const res = await utils.fetchJson(serverUrl + '/profile?address=' + encodeURIComponent(rootStoreAddress))
-      resolve(res)
-    } catch (err) {
-      reject(err)
-    }
-  })
 }
 
 module.exports = Box
